Refactor CoursePage, add course stars rating
