+++ src/components/Profile/ProfileCourses/CourseCardActions/index.tsx
@@ -17,10 +17,10 @@ import {
 } from "react";
 import { useTranslation } from "react-i18next";
 
-import { toast } from "react-toastify";
 import { ResetProgressModal } from "../ResetProgressModal";
 import { QuestionnaireModelType } from "@/types/questionnaire";
 import { Wrapper } from "./styles";
+import { getQuestionnaires } from "@/utils/questionnaires";
 
 interface Props {
   courseData: CourseProgressItem;
@@ -89,90 +89,20 @@ export const CourseCardActions: FC<Props> = ({
     }
   }, [questionnaires, state.step]);
 
-  const getQuestionnaire = useCallback(
-    async (questionnaireId: number) => {
-      try {
-        const response =
-          courseId &&
-          (await fetchQuestionnaire(
-            QuestionnaireModelType.COURSE,
-            courseId,
-            questionnaireId
-          ));
-        if (response && response.success) {
-          return response.data.questions;
-        }
-      } catch (error) {
-        toast.error(t<string>("UnexpectedError"));
-        console.log(error);
-      }
-    },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [courseId, fetchQuestionnaire]
-  );
-
-  const getQuestionnaires = useCallback(async () => {
-    setState((prevState) => ({
-      ...prevState,
-      loading: true,
-    }));
-    try {
-      const response =
-        courseData.course.id &&
-        (await fetchQuestionnaires(
-          QuestionnaireModelType.COURSE,
-          courseData.course.id
-        ));
-      if (response && response.success) {
-        const questionnairesWithCombinedQuestions = await Promise.all(
-          response.data.map(async (data) => {
-            const res = await getQuestionnaire(data.id);
-
-            const combinedQuestions = data.questions.reduce(
-              (result, element) => {
-                const matchingElement = res?.find(
-                  (item) => item.id === element.id
-                );
-
-                const updatedElement = {
-                  ...element,
-                  rate: matchingElement?.rate,
-                  note: matchingElement?.note,
-                };
-                updatedElement.public_answers === false &&
-                  updatedElement.rate === null &&
-                  updatedElement.note === null &&
-                  result.push(updatedElement);
-                return result;
-              },
-              [] as API.QuestionnaireQuestion[]
-            );
-
-            return {
-              ...data,
-              questions: combinedQuestions,
-            };
-          })
-        );
-        setQuestionnaires(
-          questionnairesWithCombinedQuestions.filter(
-            (item) => !!item.questions.length
-          )
-        );
-        setState((prevState) => ({
-          ...prevState,
-          loading: false,
-        }));
-      }
-    } catch (error) {
-      toast.error(t<string>("UnexpectedError"));
-      console.log(error);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [courseId, fetchQuestionnaires]);
-
   useEffect(() => {
-    getQuestionnaires();
+    courseId &&
+      getQuestionnaires({
+        courseId,
+        fetchQuestionnaire,
+        fetchQuestionnaires,
+        onSucces: (items) => {
+          setQuestionnaires(items);
+          setState((prevState) => ({
+            ...prevState,
+            loading: false,
+          }));
+        },
+      });
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [courseId]);
