+++ src/i18n.ts
@@ -149,6 +149,7 @@ const resources = {
         MyProgress: "My progress",
         ToSeeProgress: "to see your progress",
         UnavailableCourse: "This course is unavailable to buy",
+        NoRatings: "No ratings",
         AvarageRating: "Average rating of Jan Kamiński's training courses",
         Satisfaction: "satisfaction or your money back *",
         Guarantee: "Guaranteed",
@@ -725,6 +726,7 @@ const resources = {
         MyProgress: "Moje postępy",
         ToSeeProgress: "aby widzieć postępy",
         UnavailableCourse: "Kurs niemożliwy do kupienia",
+        NoRatings: "Brak ocen",
         AvarageRating: "Średnia ocena szkoleń Jana Kamińskiego",
         Satisfaction: "zadowolenia lub zwrot pieniędzy*",
         Guarantee: "Gwarancja",
