+++ src/pages/courses/course/Components/index.tsx
@@ -1,0 +1,13 @@
+import { CourseAuthor } from "./CourseAuthor";
+import { CourseCompanies } from "./CourseCompanies";
+import { CourseMainInfo } from "./CourseMainInfo";
+import { CourseRatings } from "./CourseRatings";
+import { CourseRelated } from "./CourseRelated";
+
+export {
+  CourseAuthor,
+  CourseCompanies,
+  CourseMainInfo,
+  CourseRatings,
+  CourseRelated,
+};
